refactor(repo): split waf.go into exclusions/snapshots files
